matchmaking/requirements.txt
django==5.0.2
channels==4.0.0
djangorestframework==3.14.0
daphne==4.1.0
websockets==12.0
aiohttp==3.9.3
pyjwt==2.8.0
bleach==6.1.0